feat(*): 拆分数据结构
